refactor(): Convert user Id type from int to long

+++ src/main/java/com/app/budget/core/services/UserService.java
@@ -67,7 +67,7 @@ public class UserService {
         jdbcUserRepository.update(updatedEntity);
     }
 
-    public void updateRole(Integer userId, String role) {
+    public void updateRole(Long userId, String role) {
         UserEntity entity = this.jdbcUserRepository.getUserById(userId);
 
         if (entity == null)
